# backend/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for external devices to access the internal Smartify system",
        "title": "Smartify Backend API",
        "contact": {
            "name": "Smartify Working Mail",
            "email": "smartifyprojectapp@gmail.com"
        },
        "version": "1.1"
    },
    "host": "213.226.112.206:22025",
    "basePath": "/api",
    "paths": {
...
        "/checkTokens": {
            "post": {
                "description": "Checks the expiration date of access and refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Checking token validity",
                "parameters": [
                    {
                        "description": "A couple of tokens to check out",
                        "name": "tokens",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Tokens_answer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The tokens are valid",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "401": {
                        "description": "Tokens are invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
        "/commit_code_reset_password": {
            "post": {
                "description": "Validates the password reset code sent to email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verifying the confirmation code",
                "parameters": [
                    {
                        "description": "Email and confirmation code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Code_verification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Code confirmed",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid code or user not found",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
        "/forgot_password": {
            "post": {
                "description": "Sends a confirmation code to the user's email for password recovery",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Password reset request",
                "parameters": [
                    {
                        "description": "User Email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Email_struct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation code sent",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid request or user not found",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
        "/hello": {
            "get": {
                "description": "Returns the status “ok” if the server is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "Checking server availability",
                "responses": {
                    "200": {
                        "description": "The server is available",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
        "/reset_password": {
            "post": {
                "description": "Sets a new password after successful verification of the confirmation code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Setting a new password",
                "parameters": [
                    {
                        "description": "Email and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Update_password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password successfully changed",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid password request or update error",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Code_verification": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "api.Email_struct": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "api.Error_answer": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "error": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
...
        "api.Success_answer": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "status": {
                    "description": "Status message",
                    "type": "string"
                }
            }
        },
        "api.Tokens_answer": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "api.Update_password": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        },
...
    }
}